numberfilter ok , textfilter testing
